requirements.txt
# ============================================================
# GEM Protocol — Python Dependencies
# ============================================================

# --- Core Framework ---
streamlit>=1.40.0
crewai>=0.80.0

# --- LLM APIs ---
google-generativeai>=0.8.0
openai>=1.50.0

# --- Data & Finance ---
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0

# --- DB ---
sqlalchemy>=2.0.0

# --- Scheduling ---
apscheduler>=3.10.0

# --- Visualization ---
plotly>=5.22.0

# --- HTTP ---
requests>=2.31.0

# --- Configuration ---
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- Utilities ---
python-dotenv>=1.0.0
openpyxl>=3.1.0
